--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "prop-types": "^15.7.2"
-  }
-}
